ai-service/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-limiter==3.7.0
groq==0.9.0
python-dotenv==1.0.1
redis==5.0.7
requests==2.32.3
pytest==8.2.2 
